+++ tests/helpers/test_utils.py
@@ -121,26 +121,3 @@ def valid_system_profile():
         "installed_services": ["ndb", "krb5"],
         "enabled_services": ["ndb", "krb5"],
     }
-
-
-def expected_headers(event_type, request_id, insights_id):
-    return {
-        "event_type": event_type,
-        "request_id": request_id,
-        "producer": os.uname().nodename,
-        "insights_id": insights_id,
-    }
-
-
-class MockEventProducer:
-    def __init__(self):
-        self.event = None
-        self.key = None
-        self.headers = None
-        self.topic = None
-
-    def write_event(self, event, key, headers, topic):
-        self.event = event
-        self.key = key
-        self.headers = headers
-        self.topic = topic
